Move CalculateHash to the Utility.cs, so that there is only one copy of it.  Call this from Single and Multiple thread code. Convert code base to single set of files with #if SQL2008 #else #endif around using statements to handle 90 to 100 changes between SQL 2005 and SQL 2008. Remove all the excess SQL 2005 files that are no longer required. Correct .Net 3.5 requirement in SQL 2005 install (changed to .Net 2.0) Add Manual Test Cases.  (There are more of these required, to cover other environments) Fix 1.1 to 1.2 upgrade issue.
If Version 1.2 is used to update a package, it will no longer run in 1.1!

Multiple Threading has now been tested, but is more expensive than Single Threading for workloads that have small numbers of output columns, and small numbers of input columns.  5 output columns of 20 input columns each (ie. 100 input columns in total) is worth using the multiple threading on...